Preparing to introduce new architecture into the root
