+++ src/architecture2/ItemsStore.ts
@@ -1,4 +1,4 @@
-import { EventsHandler } from "../domain/eventHandler";
+import { EventsHandler } from "../infra";
 
 export class ItemsStore {
   private items: Items = {};
