+++ src/architecture2/app.ts
@@ -1,9 +1,8 @@
-import { EventsHandler } from "../domain/eventHandler";
-import { renderTreeView } from "./treeView2";
+import { renderTreeView } from "./tab";
 import * as globals from "./domain";
 import { ItemsStore } from "./ItemsStore";
 import PlayerStore from "./PlayerStore";
-import { dom } from "../infra";
+import { dom, EventsHandler } from "../infra";
 import { renderPlayerFooter } from "./player";
 
 const loadLocalItems = (): Items => {
