+++ src/architecture2/tab/FolderIcon.ts
