+++ src/architecture2/tab/index.ts
