--- src/domain/index.ts
@@ -1,2 +1,0 @@
-export { events, TubeflowyEvents } from "./events";
-export * as items from "./items";
